requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
google-generativeai
numpy==1.26.4
scipy==1.12.0
pandas==2.2.2
google-genai
xmltodict==0.13.0
pymongo
flask-jwt-extended
celery
redis
bcrypt
boto3
gunicorn
flask-limiter
razorpay
docker